Special version to avoid heavy snapshot
Akka public API does not allow "controlled offset recovery"
So we store the offset oldest value as a snapshot data and
replace the metadata with it.
This allow us to not store application data in the snapshot
but recover from events from the oldest offset.
This proved to be much more efficient for our use case.
DON'T USE THIS, this broke akka-persistence specification.